requirements.txt
transformers==4.44.2
sentence-transformers==3.0.1
pypdf2==3.0.1
faiss-cpu==1.8.0.post1
streamlit==1.36.0